add connection settings
Connection settings are added at runtime while creating bridge.
This is because in the new slave connection setting,'master' key value
has to be specified which is nothing but the associated bridge for the slave.
In case of multiple ethernet interfaces bridge name (brbridgedX) depends on the
index(X) of the interface which is not constant across system restarts.

Signed-off-by: Mahantesh Salimath <salimathm@ainfosec.com>