Use system call instead of spawning a shell
'spawnShell' used to run udhcpc on eth0 interface in dom0 fails occasionally.
This leaves eth0 with no ip address.

My suspicion is that the phenomena of 'opening a shell and running a
process will result in termination of the process when the shell is closed'
is in effect when spawnShell is used here, because after few attempts
udhcpc forks a child process when it fails to get an ip address before
returning to the shell invoking it.

Using 'system' instead has no failures, system is a system call which
achieves the purpose of executing a shell command without creating a shell.

This commit also fixes the issue of having to run udhcpc on eth0 when
ethernet cable is plugged-in after system boot.

OXT-1772

Signed-off-by: Mahantesh Salimath <salimathm@ainfosec.com>