Revert "OXT-1824 [network-daemon] Close FDs when ndvm is stopped"
This reverts commit 9399c3ed281dd6393a545cb623a2d6816a3621cf.

There is an issue when ndvm's are destroyed that causes network-daemon
to close a FD but still attempt to use it after it's been closed.
This causes network-daemon to hang, and the end result is that it is
unresponsive to dbus commands until it is killed and restarted.

This was not observed when ndvm's are shutdown normally.

Until this issue is resolved, revert this commit to prevent any
regressions.