nwd: Use local MAC in transparent bridging mode
We have a case where we want to use transparent bridging and give the HW
MAC to a guest.  Before this commit, when dom0 networking is enabled,
the dom0 uses the real, HW MAC address, and NDVM changes the NIC's MAC
to have the local bit set.  If dom0 networking is disabled, the NDVM
keeps the real HW MAC, so it cannot be assigned to a guest.

Expand the configureNetwork code, so that dom0-networking enabled and
transparent bridging are handled identically.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>